Fix cache condition, refactor caching
