+++ gramps_webapi/api/resources/relations.py
@@ -30,12 +30,11 @@ from webargs import fields, validate
 from gramps_webapi.api.people_families_cache import CachePeopleFamiliesProxy
 
 from ...types import Handle
-from ..cache import request_cache
+from ..cache import request_cache_decorator
 from ..util import (
     abort_with_message,
     get_db_handle,
     get_locale_for_language,
-    make_cache_key_request,
     use_args,
 )
 from . import ProtectedResource
@@ -55,7 +54,7 @@ class RelationResource(ProtectedResource, GrampsJSONEncoder):
         },
         location="query",
     )
-    @request_cache.cached(make_cache_key=make_cache_key_request)
+    @request_cache_decorator
     def get(self, args: Dict, handle1: Handle, handle2: Handle) -> Response:
         """Get the most direct relationship between two people."""
         db_handle = CachePeopleFamiliesProxy(get_db_handle())
@@ -101,7 +100,7 @@ class RelationsResource(ProtectedResource, GrampsJSONEncoder):
         },
         location="query",
     )
-    @request_cache.cached(make_cache_key=make_cache_key_request)
+    @request_cache_decorator
     def get(self, args: Dict, handle1: Handle, handle2: Handle) -> Response:
         """Get all possible relationships between two people."""
         db_handle = CachePeopleFamiliesProxy(get_db_handle())
